+++ src/pages/components/tabBarApis/tabBarApis.scss
@@ -1,4 +1,4 @@
-@import "../../../../../styles/variables.scss";
+@import "../../../styles/variables.scss";
 
 .btn {
   margin-bottom: 40px;
